after Lecture #3
typed all code in lecture #3, code is now ready for the assignment.
but still I have to fix the scoring, as its messed up some how. score
is not going in positive anytime and its being jumped minus 3, instead
of minus 1, when I am clicking, different card which are not matching